Translated the 'z' cone constraint to 'f' in format_and_copy_cone(cone) function
SCS's core programs in C consider 'f' as the number of linear equality constraints (primal zero, dual free) even though the new version on the website reflects this as 'z'. Turns out they both are the same cone constraints. So, I made this modification so that if we get 'z' it is still processed as 'f'.